fix: restructure test dir
