+++ pytest.ini
@@ -8,4 +8,3 @@ log_cli_format=[%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
 # Uncomment the following if you need timestamps
 # log_cli_format=%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
 log_cli_date_format=%Y-%m-%d %H:%M:%S
-
