+++ tests/osman/conftest.py
@@ -1,45 +1,58 @@
 """
-Tests initialization and setup
+Osman instance test configuration and setup
 """
+import time
 import logging
-import os
-import sys
 import pytest
 
-# This ugly hack allows importing 'osman.environment' without initializing
-# osman package (i.e. without executing osman/__init__.py)
-# We need to prevent initializing OsmanConfig class
-ORIG_SYS_PATH=sys.path.copy()
-sys.path.append(os.path.join(".", "osman"))
-from environment import OSMAN_ENVIRONMENT_VARS
-sys.path = ORIG_SYS_PATH
-
-# Dictionary to save deleted environment variables
-OSMAN_ENV_VARS_SAVED = {}
-def pytest_configure():
+def get_random_index_name(prefix: str) -> str:
     """
-    Propagate global variable to the tests.
+    Return a random index name
     """
-    pytest.OSMAN_ENV_VARS_SAVED = OSMAN_ENV_VARS_SAVED
-
-def save_and_delete_osman_environment_vars():
-    """
-    Delete Osman's variables from the environment.
-    """
-    logging.info("Deleting Osman environment variables")
-    for variable in OSMAN_ENVIRONMENT_VARS + ["AWS_USER", "AWS_SECRET"]:
-        if os.environ.get(variable) is None:
-            continue
-        OSMAN_ENV_VARS_SAVED[variable] = os.environ.pop(variable)
+    return f"{prefix}_{int(time.time()*1_000_000)}"
 
 @pytest.fixture
-def save_and_delete_osman_environment_vars_fixture():
+def random_index_name(request: pytest.FixtureRequest):
     """
-    Pytest fixture, workaround as fixtures can's be called directly.
+    Fixture for getting random index name. Usefull when testing
+    manual index manipulation
     """
-    save_and_delete_osman_environment_vars()
+    return get_random_index_name(request.function.__name__)
+
+@pytest.fixture
+def index_handler(request: pytest.FixtureRequest):
+    """
+    Creates new index, yields index name to test case so it can use it
+    and after the test is done deletes the index.
+    Expects the following params in 'request':
+    os_man: Osman
+        Osman instance
+    mapping: dict
+        Index mapping, can be None
+    """
+    assert request.param
+    assert request.param["os_man"]
+    assert request.param["mapping"]
+
+    # Get name of the caller function (name of the test)
+    caller_name = request.function.__name__
+
+    # Create index name as `caller_name` + timestamp
+    index_name = get_random_index_name(caller_name)
+    logging.info("Creating index '%s'", index_name)
+
+    mapping = request.param["mapping"]
+    os_man = request.param["os_man"]
+
+    # Create new index with `index_name` and optional mapping using the helper
+    res = os_man.create_index(index_name, mapping)
+    logging.debug(res)
+
+    # Return name of the index to the test that is using this fixture
+    yield index_name
+
+    # Once the test ends (or crash) this line will be executed
+    os_man.delete_index(index_name)
+
 
 logging.info("Starting: '%s'", __file__)
-# This can't be done through fixtures, it has to be called
-# before any 'import osman'
-save_and_delete_osman_environment_vars()
